expense choose action and refactor
